+++ lib/librte_ip_frag/rte_ip_frag.h
@@ -147,6 +147,23 @@ struct rte_ip_frag_tbl * rte_ip_frag_table_create(uint32_t bucket_num,
 		uint64_t max_cycles, int socket_id);
 
 /**
+ * Sweep the IP fragmentation table for expired segments.
+ *
+ * @param tbl
+ *   Fragmentation table to sweep.
+ * @param dr
+ *   Death row to free buffers to
+ * @param next
+ *   Pointer to the bucket iterator.
+ *   Should be 0 to start sweeping the fragmentation table.
+ *   Bucket iterator is incremented after each call of this function.
+ * @return
+ *   0 if successful. -EINVAL if the parameters are invalid.
+ */
+int rte_ip_frag_sweep_table(struct rte_ip_frag_tbl *tbl,
+	struct rte_ip_frag_death_row *dr, uint32_t *next);
+
+/**
  * Free allocated IP fragmentation table.
  *
  * @param tbl
